+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### dev-master
+
+* Fix warning for `crypt()` and salt in PHP7
+* Travis
+
 ### 0.1.2 (03.12.2015)
 
 * #3: Fix `getContent()` for empty file